Fixed code readability issues Updated serializes folder structure
