+++ Assets/Scripts/Runtime/Serializers/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad3fc2b8175d6034b863d1e0dbe8185f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
