+++ Assets/Scripts/Runtime/Serializers/Inventory/AccessoryItem.cs
