+++ Assets/Scripts/Runtime/Serializers/Inventory/ConsumableItem.cs
