+++ Assets/Scripts/Runtime/Serializers/Inventory/InventoryItem.cs
