+++ Assets/Scripts/Runtime/Serializers/Inventory/RelicItem.cs
