+++ Assets/Scripts/Runtime/Serializers/Party.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36f378a778e05864c979979d38ae830b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
