+++ Assets/Scripts/Runtime/Serializers/Party/PartyMember.cs.meta
