+++ Assets/Scripts/Runtime/Serializers/PlayerPrefs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f63902a7e615b324c98f757de6114a93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
