+++ Assets/Scripts/Runtime/Serializers/Save.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77c01027f04fb4e4ba5b494e1e31545a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
